% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsbdtemplate.R
\docType{package}
\name{dsbdtemplate-package}
\alias{dsbdtemplate}
\alias{dsbdtemplate-package}
\title{dsbdtemplate: Rmarkdown Theme for DSBD documents}
\description{
A set of custom Rmarkdown templates for documents and presentations.
}
\details{
Trial run on RMarkdown templates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fernandomayer/dsbdtemplate}
  \item Report bugs at \url{https://github.com/fernandomayer/dsbdtemplate/issues}
}

}
\author{
\strong{Maintainer}: Fernando Mayer \email{fernando.mayer@ufpr.br}

}
